More consistency fixes. Added business widgets UI

+++ AdvantClickerTT/Assets/Scripts/ECS/Components/GameTimeComponent.cs
@@ -1,4 +1,4 @@
 public struct GameTimeComponent
 {
-    public double CurrentTime;
+    public float CurrentTime;
 }
